[IMP] website: extracted language selector in its own template
It is a technical part of the website, it actually acts like a "Language Widget", I don't see any good reason to not extract it into its own template.
I needed this for a designer who wanted to move the language selection elsewhere, and I realized it was not already in its own template. Worse, we already had a duplication of this content in our own module!
I need this in saas-6 to get rid of the actual third duplications of this content which exists in 8.0 (counting the theme).