[FIX] mrp: Extra moves and multiple split corrections in consumed and produced moves
ORM will show lines in production.move_lines even if done (which the domain should avoid)
when processing the production (as they are probably still in cache).  We explicitly skip those lines.

When producing more with lots in consumed and produced, we added the consumed_for link as
the lot tracking requires to have this link between the consumed and produced moves.

The extra move of the produced goods should also be confirmed and not created as confirmed immediately.
That way it can still apply the push rules.  (Gabriel Demmerle)