[FIX] web: prevent graph view from crashing (sometimes)
if the graph view is set in bar mode, and does not have any groupby,
then it tried to read some property from its viewmanager, which results
in a crash.  It is most likely the remnants of the old code, this commit
correctly sets the string to the current measure.