[FIX] Web: view_list.js: _.template()
The signature and behavior of function template() of underscore.js has
been changed from rev. 1.7.0. It doesn't accept an initial data object
anymore, and always returns a function.

This commits adapts our calls to _.template(). This fixes display bugs in
list views (e.g. the progressbar which was not displayed).