[FIX] stock: UoM of a product should not be changed when done stock moves exist.
The UoM of a product can be changed to whatever UoM if there have not been made any done stock moves.
If there are, you can not change the UoM.  If there are not done moves, you can not change the category
anymore as the conversion of the product_qty will give an error.

The reason we put the constraint is that we would need to change all the qty values of the quants
when we change UoM.  It is better to avoid this (e.g. rounding).  That is why we put the constraint.

There is no reason to restrict change of uom_po_id, except that it should
fall within the same category as the UoM.  (constraint)