[FIX] Web: view.js: rpc call to load sidebar content
this.options.$sidebar usage has been changed at rev. a5f68b3b28f6da156ad88e1c9e3b6de955471441
and a flag this.options.sidebar was introduced ($sidebar now simply
represents a jQuery element where the sidebar has to be rendered,
whereas sidebar represents whether or not a sidebar has to be rendered).

The old variable $sidebar was still used for the rpc to get the content
of the sidebar stored server-side, resulting in a sidebar with missing
content.