[FIX] base: update of source value in translation
Make sure pass a list of ids instead of a single id to write calls.
Some models (e.g. blog.post, fixed at 1cd636f) are assuming that it got a list
of ids and is not checking the type.
Always pass a list of ids as it's the expected format for the orm.