[FIX] mrp: ability to search BOM on BOM lines containing ...
In BOM, when performing an advanced search
on "BOM Lines" contains "a name"
all lines were returned, whatever the lines content.

This was due to the simple fact no field 'name'
was set on the mrp.bom.line model.

We set "product_id" as _rec_name, it seems the more
logical choice.

opw-631335