[FIX] customer: sale order number with child
The number of sale order on a partner form view only counted the given customer
and not the ones of eventuals child_ids. Hence when clicking on the widget, the
result sale order tree view (which take account of child_ids sale order) is
inconsistent.

This fix adds the sale order of a child company to the sale_order_count.

opw-630249