[FIX] hr_timesheet_sheet: timesheet report totals
This rev. is related to rev. abb21a2

If several account analytic lines where defined for a same date
within a same sheet, the report totals were multiplied by
the number of occurences of lines of the same day.

Before the above rev., only `Diff`, `Total`, `Timesheet`, were wrong
After, other totals were wrong as well.

The point of the above rev. was to gain a real performance improvment
when there was a lot of timesheets / analytic lines entered
in database. This rev. keeps this performance improvment.

opw-629857