[FIX] stock_account: avoid timezone periode mismatch
The `find` method will get the current accounting periode according to the
user's timezone while time do not. This means that it is possible that users
get a move with a date belongign to a different period that the one returned by
`find`.
Fixing 6f127fe, linked to #4147, courtesy of Graeme Gellatly