[FIX] im_livechat : remove async attribute for internal livechat lib
The internal livechat loader should not be loading in a asynchronous way, since it depends on the other assets bundle included in a odoo website page. Sometime (depending on the browser cache), this triggered a 'TypeError' when executing 'new openerp.im_livechat.LiveSupport', since 'openerp' was 'undefined'. This was introduced with 21298a1e6a79c030601832ece8916e9d72f0ae29.