[FIX] website_blog: don't add background from backend
Adding an image from the frontend use ir.attachment or pre-existing url
and set the url in the blog.post's background_image field. This is
not the same as a binary field which sets the data in the field.

Since it's possible to do it in the frontend, there is no use of the
field in the backend. This fix removes this field.

closes #6497
issue #1305
opw-633725